requirements.txt
discord.py==2.3.2
anthropic==0.40.0
boto3==1.35.0
botocore==1.35.0
python-dotenv==1.0.1
aiohttp==3.10.5
PyNaCl==1.5.0
